+++ api/middleware/inbound.go
@@ -26,10 +26,10 @@ import (
 	"go.uber.org/yarpc/api/transport"
 )
 
-// UnaryInboundMiddleware defines a transport-level middleware for
+// UnaryInbound defines a transport-level middleware for
 // `UnaryHandler`s.
 //
-// UnaryInboundMiddleware MAY
+// UnaryInbound middleware MAY
 //
 // - change the context
 // - change the request
@@ -38,99 +38,99 @@ import (
 // - handle the returned error
 // - call the given handler zero or more times
 //
-// UnaryInboundMiddleware MUST be thread-safe.
+// UnaryInbound middleware MUST be thread-safe.
 //
-// UnaryInboundMiddleware is re-used across requests and MAY be called multiple times
+// UnaryInbound middleware is re-used across requests and MAY be called multiple times
 // for the same request.
-type UnaryInboundMiddleware interface {
+type UnaryInbound interface {
 	Handle(ctx context.Context, req *transport.Request, resw transport.ResponseWriter, h transport.UnaryHandler) error
 }
 
-// NopUnaryInboundMiddleware is a inbound middleware that does not do anything special. It
+// NopUnaryInbound is a inbound middleware that does not do anything special. It
 // simply calls the underlying Handler.
-var NopUnaryInboundMiddleware UnaryInboundMiddleware = nopUnaryInboundMiddleware{}
+var NopUnaryInbound UnaryInbound = nopUnaryInbound{}
 
-// ApplyUnaryInboundMiddleware applies the given InboundMiddleware to the given Handler.
-func ApplyUnaryInboundMiddleware(h transport.UnaryHandler, i UnaryInboundMiddleware) transport.UnaryHandler {
+// ApplyUnaryInbound applies the given InboundMiddleware to the given Handler.
+func ApplyUnaryInbound(h transport.UnaryHandler, i UnaryInbound) transport.UnaryHandler {
 	if i == nil {
 		return h
 	}
 	return unaryHandlerWithMiddleware{h: h, i: i}
 }
 
-// UnaryInboundMiddlewareFunc adapts a function into an InboundMiddleware.
-type UnaryInboundMiddlewareFunc func(context.Context, *transport.Request, transport.ResponseWriter, transport.UnaryHandler) error
+// UnaryInboundFunc adapts a function into an InboundMiddleware.
+type UnaryInboundFunc func(context.Context, *transport.Request, transport.ResponseWriter, transport.UnaryHandler) error
 
-// Handle for UnaryInboundMiddlewareFunc
-func (f UnaryInboundMiddlewareFunc) Handle(ctx context.Context, req *transport.Request, resw transport.ResponseWriter, h transport.UnaryHandler) error {
+// Handle for UnaryInboundFunc
+func (f UnaryInboundFunc) Handle(ctx context.Context, req *transport.Request, resw transport.ResponseWriter, h transport.UnaryHandler) error {
 	return f(ctx, req, resw, h)
 }
 
 type unaryHandlerWithMiddleware struct {
 	h transport.UnaryHandler
-	i UnaryInboundMiddleware
+	i UnaryInbound
 }
 
 func (h unaryHandlerWithMiddleware) Handle(ctx context.Context, req *transport.Request, resw transport.ResponseWriter) error {
 	return h.i.Handle(ctx, req, resw, h.h)
 }
 
-type nopUnaryInboundMiddleware struct{}
+type nopUnaryInbound struct{}
 
-func (nopUnaryInboundMiddleware) Handle(ctx context.Context, req *transport.Request, resw transport.ResponseWriter, handler transport.UnaryHandler) error {
+func (nopUnaryInbound) Handle(ctx context.Context, req *transport.Request, resw transport.ResponseWriter, handler transport.UnaryHandler) error {
 	return handler.Handle(ctx, req, resw)
 }
 
-// OnewayInboundMiddleware defines a transport-level middleware for
+// OnewayInbound defines a transport-level middleware for
 // `OnewayHandler`s.
 //
-// OnewayInboundMiddleware MAY
+// OnewayInbound middleware MAY
 //
 // - change the context
 // - change the request
 // - handle the returned error
 // - call the given handler zero or more times
 //
-// OnewayInboundMiddleware MUST be thread-safe.
+// OnewayInbound middleware MUST be thread-safe.
 //
-// OnewayInboundMiddleware is re-used across requests and MAY be called
+// OnewayInbound middleware is re-used across requests and MAY be called
 // multiple times for the same request.
-type OnewayInboundMiddleware interface {
+type OnewayInbound interface {
 	HandleOneway(ctx context.Context, req *transport.Request, h transport.OnewayHandler) error
 }
 
-// NopOnewayInboundMiddleware is an inbound middleware that does not do
+// NopOnewayInbound is an inbound middleware that does not do
 // anything special. It simply calls the underlying OnewayHandler.
-var NopOnewayInboundMiddleware OnewayInboundMiddleware = nopOnewayInboundMiddleware{}
+var NopOnewayInbound OnewayInbound = nopOnewayInbound{}
 
-// ApplyOnewayInboundMiddleware applies the given OnewayInboundMiddleware to
+// ApplyOnewayInbound applies the given OnewayInbound middleware to
 // the given OnewayHandler.
-func ApplyOnewayInboundMiddleware(h transport.OnewayHandler, i OnewayInboundMiddleware) transport.OnewayHandler {
+func ApplyOnewayInbound(h transport.OnewayHandler, i OnewayInbound) transport.OnewayHandler {
 	if i == nil {
 		return h
 	}
 	return onewayHandlerWithMiddleware{h: h, i: i}
 }
 
-// OnewayInboundMiddlewareFunc adapts a function into a OnwayInboundMiddleware.
-type OnewayInboundMiddlewareFunc func(context.Context, *transport.Request, transport.OnewayHandler) error
+// OnewayInboundFunc adapts a function into a OnewayInbound Middleware.
+type OnewayInboundFunc func(context.Context, *transport.Request, transport.OnewayHandler) error
 
-// HandleOneway for OnewayInboundMiddlewareFunc
-func (f OnewayInboundMiddlewareFunc) HandleOneway(ctx context.Context, req *transport.Request, h transport.OnewayHandler) error {
+// HandleOneway for OnewayInboundFunc
+func (f OnewayInboundFunc) HandleOneway(ctx context.Context, req *transport.Request, h transport.OnewayHandler) error {
 	return f(ctx, req, h)
 }
 
 type onewayHandlerWithMiddleware struct {
 	h transport.OnewayHandler
-	i OnewayInboundMiddleware
+	i OnewayInbound
 }
 
 func (h onewayHandlerWithMiddleware) HandleOneway(ctx context.Context, req *transport.Request) error {
 	return h.i.HandleOneway(ctx, req, h.h)
 }
 
-type nopOnewayInboundMiddleware struct{}
+type nopOnewayInbound struct{}
 
-func (nopOnewayInboundMiddleware) HandleOneway(ctx context.Context, req *transport.Request, handler transport.OnewayHandler) error {
+func (nopOnewayInbound) HandleOneway(ctx context.Context, req *transport.Request, handler transport.OnewayHandler) error {
 	return handler.HandleOneway(ctx, req)
 }